defaults: Move defaults into pkg/defaults
This helps resolve import cycle problems for future commits
