+++ cilium/cmd/cleanup.go
@@ -22,8 +22,8 @@ import (
 	"strings"
 
 	"github.com/cilium/cilium/common"
-	"github.com/cilium/cilium/daemon/defaults"
 	"github.com/cilium/cilium/pkg/bpf"
+	"github.com/cilium/cilium/pkg/defaults"
 	"github.com/cilium/cilium/pkg/maps/tunnel"
 
 	"github.com/spf13/cobra"
